+++ src/emulator/extensionsEmulator.ts
@@ -10,14 +10,8 @@ import { downloadExtensionVersion } from "./download";
 import { EmulatableBackend } from "./functionsEmulator";
 import { getExtensionFunctionInfo } from "../extensions/emulator/optionsHelper";
 import { EmulatorLogger } from "./emulatorLogger";
+import { Emulators } from "./types";
 
-const EMULATED_APIS = [
-  "storage-component.googleapis.com",
-  "firestore.googleapis.com",
-  "pubsub.googleapis.com",
-  "identitytoolkit.googleapis.com",
-  // TODO: Is there a RTDB API we need to add here? I couldn't find one.
-];
 export interface ExtensionEmulatorArgs {
   projectId: string;
   projectNumber: string;
@@ -31,6 +25,7 @@ export interface ExtensionEmulatorArgs {
 export class ExtensionsEmulator {
   private want: planner.InstanceSpec[] = [];
   private args: ExtensionEmulatorArgs;
+  private logger = EmulatorLogger.forEmulator(Emulators.EXTENSIONS);
 
   constructor(args: ExtensionEmulatorArgs) {
     this.args = args;
@@ -172,19 +167,5 @@ export class ExtensionsEmulator {
       DATABASE_URL: `https://${projectId}.firebaseio.com`,
       STORAGE_BUCKET: `${projectId}.appspot.com`,
     };
-  }
-
-  private async checkAndWarnAPIs(instances: planner.InstanceSpec[]): Promise<void> {
-    const apisInstancesMap: Record<string, string[]> = {};
-    for (const i of instances) {
-      const extensionVersion = await planner.getExtensionVersion(i);
-      for (const api of extensionVersion.spec.apis ?? []) {
-        if (apisInstancesMap[api.apiName]) {
-          apisInstancesMap[api.apiName].push(i.instanceId);
-        } else {
-          apisInstancesMap[api.apiName] = [i.instanceId];
-        }
-      }
-    }
   }
 }
